Use `DESTDIR` to create binary archives
We used to do a trick setting the date of installed files to 1980 to
identify files installed by a specific component. This lead to
problems. This commit drops the previous approach in favor of employing
the `DESTDIR` environment variable which is commonly used to choose the
install location to use in front of the prefix.

Boost and Linux headers don't have explicit support for `DESTDIR`,
therefore small workarounds were necessary.