Fix build of alltypes.h for newer versions of musl
In recent versions of musl the header alltypes.h has been moved to a new
path in the project. This breaks the build because the targets in the
Makefile need to take into account the new path. With this commit we try
both the old and the new targets, so that we can build it correctly
independently of the version of musl.