Fix libc dependencies
libcs used to depend on `gcc-stage1`, but this was introducing a
circular dependency. In fact, since we now remove files from
`installed-targets/` upon install, when installing `gcc-stage2`, the
corresponding `gcc-stage1` file was deleted. But `gcc-stage2` depends on
the libc which, in turn, depends on `gcc-stage1`.

The libc now depends on `gcc` and `gcc-stage1` is the default build for
`gcc`. This solves the problem.