Fix build for musl version >= 1.1.15
Since version 1.1.15 the musl buildsystem introduced a mechanism in the
configure script for autodetecting cross-compilation and setting the
CROSS_COMPILE environment variable accordingly if it was not set by the
caller.

This mechanism broke our build because:
1) we do not set CROSS_COMPILE even if we're cross compiling, so the new
musl configure sets CROSS_COMPILE automatically to the correct triple;
2) before this commit we didn't have in the gcc install directory and
the binutils install directory in PATH when compiling musl, so the build
system was not able to detect the proper build tools for cross
compilation.

Now we source the environment before building musl, so that we're sure
that the build tools with the proper cross-compilation triple are found
in the PATH.