Fix GCC and libc dependencies
This commit tries (again) to fix dependencies issues that led to a
non-stable loop of builds when giving multiple times the same
command. The reason for this was that, for instance, `gcc-stage2`
installed `gcc`, which should be fine according to the latest attempt to
fix the problem, except for the fact that this means touching the file
`installed-targets/toolchain/*/gcc`, making it look newer and therefore
triggering rebuilds.

The solution is to introduce more explicitly the fact that a certain
build "provides" another build. For instance, `gcc-stage2` "provides"
`gcc-stage1` and `libc-default` "provides" `libc-headers`. A build that
provides another build will touch the corresponding install file, unless
it already exists. If it already exists, the file is left untouched,
preserving the dependencies graph in the right form.

In addition, dependencies are now separated between "configure-time" and
"install-time". Install-time are always required, while configure-time
dependencies are required only if we're building from source.

* The concept of providing another build is represented by the
  `_PROVIDES` build variable.
* The distinction between configure-time and install-time dependencies
  is expressed by differentiating the `_CONFIGURE_DEPS` and
  `_INSTALL_DEPS` build variables.
* For improved readibility, this commit now splits `$(calls)` onto
  multiple lines. However, when this is done, the call has to be wrapped
  in another function that strips from whitespace each argument.
* In `toolchain.mk`, the order in which the libc and GCC components are
  generated has been changed, so that we can correctly depend on `gcc`
  being installed before building the libc and `gcc-stage2` can depend
  on libc being installed before building.