Build binary archives only if needed
In the past we used to rebuild all the binary archives, always, no
matter what. This commit rebuilds a build only if the repository
associated to that build has a commit not available in the binary
archive or if the files that generated the corresponding component
changed.

For this reason, we also split Makefile in several smaller files we can
invoke `git log` on. In fact, to verify if the file that generated a
component has changed, we check the last commit affecting it.