Patch boost to prevent GCC 8 warnings
GCC 8 adds stricter tests for `-Wparentheses` warnings.  These warnings,
that happen to be triggered by the boost version we employ, cause
compilation failures when `revamb` is built with `-Werror` (like it's
mostly common in development).

I didn't want to remove or silence the warnings in `revamb`, since they
are sometimes useful to spot real bugs, so I backported here a patch
coming from upstream boost, which disables only this specific warning
and only for the interesting points where the boost developers have
already decided that it's not harmful to suppress these warnings.