Decouple toolchains list from `toolchain.mk`
`toolchain.mk` used to contain the list of toolchains to build. Since
each toolchain uses (among other things) the last commit of
`toolchain.mk` to detect if a rebuild is required, this forced us to
rebuild all the toolchains each time a new toolchain is introduced.

This commit moves in a standalone file (`toolchains-list.mk`) the
toolchain list.