Import `llvmcpy`
This commit imports `llvmcpy`, our Python library for interacting with
LLVM IR. To do so, the generation of the `environment` script has been
extended to export `PYTHONPATH`, so that `orchestra` users can easily
use our eggs.

To implement Python support in `orchestra`, we now provide the source
path to build and install stages too.